v4.0.0
Expect breaking changes everywhere, this is almost a rewrite. See `GroupVoiceCallMirrorSample.cs` on new usage.

Major improvements:
- support for audio filters
- server side muting and deafening
- Separate interfaces for audio server and client
- Opus encoding and decoding

UniMic, UnityOpus and BRW are now default dependencies of the project.

Mirror implementation and sample are included. To activate the Mirror code, add Mirror to the project and add the `UNIVOICE_MIRROR_NETWORK` compilation symbol to the project.